Move with 2 sets of animations and speed control
Still in the Love engine. This file has "perfectly" working animation
controls after about 8 hours of coding past the last sync. The idea is
that not only will it remember which direction the player was facing,
but by holding "x" you can switch the animations at anytime from a walk
to a run cycle. You can also release the "x" at any time to revert it.
This brings me that much closer to the walking controls I want. All
that's left is grid movement and the ability to turn the player without
moving using a quick direction press.