Love: Move Code
Final is a bit of a misnomer. This code is super long and needs
trimming, but it has everything I wanted. Two sets of smooth looking
gridlock animations. the ability to change speeds without unsnapping
from the grid and turning without moving but quickly pressing a
direction. In essence, it's GBA pokemon movement.